feat: update unauthenticated navbar
